+++ inc/utils/carbon_memory.h
@@ -1,3 +1,3 @@
 #pragma once
 #include "utils/carbon_commons.h"
-void *carbon_reallocate(size_t oldSize, size_t newSize, void *oldptr);
+void *carbon_reallocate(uint32_t oldSize, uint32_t newSize, void *oldptr);
